+++ packages/adapter-vercel/index.js
@@ -5,6 +5,7 @@ import { nodeFileTrace } from '@vercel/nft';
 import esbuild from 'esbuild';
 import { get_pathname } from './utils.js';
 
+const name = '@sveltejs/adapter-vercel';
 const DEFAULT_FUNCTION_NAME = 'fn';
 
 const get_default_runtime = () => {
@@ -17,6 +18,9 @@ const get_default_runtime = () => {
 	);
 };
 
+// https://vercel.com/docs/functions/edge-functions/edge-runtime#compatible-node.js-modules
+const compatible_node_modules = ['async_hooks', 'events', 'buffer', 'assert', 'util'];
+
 /** @type {import('.').default} **/
 const plugin = function (defaults = {}) {
 	if ('edge' in defaults) {
@@ -24,7 +28,7 @@ const plugin = function (defaults = {}) {
 	}
 
 	return {
-		name: '@sveltejs/adapter-vercel',
+		name,
 
 		async adapt(builder) {
 			if (!builder.routes) {
@@ -59,10 +63,12 @@ const plugin = function (defaults = {}) {
 
 			/**
 			 * @param {string} name
-			 * @param {import('./index.js').ServerlessConfig} config
-			 * @param {import('@sveltejs/kit').RouteDefinition<import('./index.js').Config>[]} routes
+			 * @param {import('.').ServerlessConfig} config
+			 * @param {import('@sveltejs/kit').RouteDefinition<import('.').Config>[]} routes
 			 */
 			async function generate_serverless_function(name, config, routes) {
+				const dir = `${dirs.functions}/${name}.func`;
+
 				const relativePath = path.posix.relative(tmp, builder.getServerDirectory());
 
 				builder.copy(`${files}/serverless.js`, `${tmp}/index.js`, {
@@ -77,19 +83,18 @@ const plugin = function (defaults = {}) {
 					`export const manifest = ${builder.generateManifest({ relativePath, routes })};\n`
 				);
 
-				await create_function_bundle(
-					builder,
-					`${tmp}/index.js`,
-					`${dirs.functions}/${name}.func`,
-					config,
-					routes
-				);
+				await create_function_bundle(builder, `${tmp}/index.js`, dir, config);
+
+				for (const asset of builder.findServerAssets(routes)) {
+					// TODO use symlinks, once Build Output API supports doing so
+					builder.copy(`${builder.getServerDirectory()}/${asset}`, `${dir}/${asset}`);
+				}
 			}
 
 			/**
 			 * @param {string} name
-			 * @param {import('./index.js').EdgeConfig} config
-			 * @param {import('@sveltejs/kit').RouteDefinition<import('./index.js').EdgeConfig>[]} routes
+			 * @param {import('.').EdgeConfig} config
+			 * @param {import('@sveltejs/kit').RouteDefinition<import('.').EdgeConfig>[]} routes
 			 */
 			async function generate_edge_function(name, config, routes) {
 				const tmp = builder.getBuildDirectory(`vercel-tmp/${name}`);
@@ -107,20 +112,61 @@ const plugin = function (defaults = {}) {
 					`export const manifest = ${builder.generateManifest({ relativePath, routes })};\n`
 				);
 
-				await esbuild.build({
-					entryPoints: [`${tmp}/edge.js`],
-					outfile: `${dirs.functions}/${name}.func/index.js`,
-					target: 'es2020', // TODO verify what the edge runtime supports
-					bundle: true,
-					platform: 'browser',
-					format: 'esm',
-					external: config.external,
-					sourcemap: 'linked',
-					banner: { js: 'globalThis.global = globalThis;' },
-					loader: {
-						'.wasm': 'copy'
+				try {
+					const result = await esbuild.build({
+						entryPoints: [`${tmp}/edge.js`],
+						outfile: `${dirs.functions}/${name}.func/index.js`,
+						target: 'es2020', // TODO verify what the edge runtime supports
+						bundle: true,
+						platform: 'browser',
+						format: 'esm',
+						external: [
+							...compatible_node_modules,
+							...compatible_node_modules.map((id) => `node:${id}`),
+							...(config.external || [])
+						],
+						sourcemap: 'linked',
+						banner: { js: 'globalThis.global = globalThis;' },
+						loader: {
+							'.wasm': 'copy'
+						}
+					});
+
+					if (result.warnings.length > 0) {
+						const formatted = await esbuild.formatMessages(result.warnings, {
+							kind: 'warning',
+							color: true
+						});
+
+						console.error(formatted.join('\n'));
 					}
-				});
+				} catch (error) {
+					for (const e of error.errors) {
+						for (const node of e.notes) {
+							const match =
+								/The package "(.+)" wasn't found on the file system but is built into node/.exec(
+									node.text
+								);
+
+							if (match) {
+								node.text = `Cannot use "${match[1]}" when deploying to Vercel Edge Functions.`;
+							}
+						}
+					}
+
+					const formatted = await esbuild.formatMessages(error.errors, {
+						kind: 'error',
+						color: true
+					});
+
+					console.error(formatted.join('\n'));
+
+					throw new Error(
+						`Bundling with esbuild failed with ${error.errors.length} ${
+							error.errors.length === 1 ? 'error' : 'errors'
+						}`
+					);
+				}
 
 				write(
 					`${dirs.functions}/${name}.func/.vc-config.json`,
@@ -136,7 +182,7 @@ const plugin = function (defaults = {}) {
 				);
 			}
 
-			/** @type {Map<string, { i: number, config: import('./index.js').Config, routes: import('@sveltejs/kit').RouteDefinition<import('./index.js').Config>[] }>} */
+			/** @type {Map<string, { i: number, config: import('.').Config, routes: import('@sveltejs/kit').RouteDefinition<import('.').Config>[] }>} */
 			const groups = new Map();
 
 			/** @type {Map<string, { hash: string, route_id: string }>} */
@@ -145,7 +191,7 @@ const plugin = function (defaults = {}) {
 			/** @type {Map<string, string>} */
 			const functions = new Map();
 
-			/** @type {Map<import('@sveltejs/kit').RouteDefinition<import('./index.js').Config>, { expiration: number | false, bypassToken: string | undefined, allowQuery: string[], group: number, passQuery: true }>} */
+			/** @type {Map<import('@sveltejs/kit').RouteDefinition<import('.').Config>, { expiration: number | false, bypassToken: string | undefined, allowQuery: string[], group: number, passQuery: true }>} */
 			const isr_config = new Map();
 
 			/** @type {Set<string>} */
@@ -164,7 +210,7 @@ const plugin = function (defaults = {}) {
 				}
 
 				const node_runtime = /nodejs([0-9]+)\.x/.exec(runtime);
-				if (runtime !== 'edge' && (!node_runtime || +node_runtime[1] < 18)) {
+				if (runtime !== 'edge' && (!node_runtime || node_runtime[1] < 18)) {
 					throw new Error(
 						`Invalid runtime '${runtime}' for route ${route.id}. Valid runtimes are 'edge' and 'nodejs18.x' or higher ` +
 							'(see the Node.js Version section in your Vercel project settings for info on the currently supported versions).'
@@ -336,6 +382,21 @@ const plugin = function (defaults = {}) {
 			builder.log.minor('Writing routes...');
 
 			write(`${dir}/config.json`, JSON.stringify(static_config, null, '\t'));
+		},
+
+		supports: {
+			// reading from the filesystem only works in serverless functions
+			read: ({ config, route }) => {
+				const runtime = config.runtime ?? defaults.runtime;
+
+				if (runtime === 'edge') {
+					throw new Error(
+						`${name}: Cannot use \`read\` from \`$app/server\` in route \`${route.id}\` configured with \`runtime: 'edge'\``
+					);
+				}
+
+				return true;
+			}
 		}
 	};
 };
@@ -369,7 +430,7 @@ function write(file, data) {
 // This function is duplicated in adapter-static
 /**
  * @param {import('@sveltejs/kit').Builder} builder
- * @param {import('index.js').Config} config
+ * @param {import('.').Config} config
  */
 function static_vercel_config(builder, config) {
 	/** @type {any[]} */
@@ -378,11 +439,8 @@ function static_vercel_config(builder, config) {
 	/** @type {Record<string, { path: string }>} */
 	const overrides = {};
 
-	/** @type {import('./index.js').ImagesConfig | undefined} */
-	let images;
-	if (config.runtime !== 'edge') {
-		images = /** @type {import('./index.js').ServerlessConfig} */ (config).images;
-	}
+	/** @type {import('./index').ImagesConfig} */
+	const images = config.images;
 
 	for (const [src, redirect] of builder.prerendered.redirects) {
 		prerendered_redirects.push({
@@ -438,10 +496,9 @@ function static_vercel_config(builder, config) {
  * @param {import('@sveltejs/kit').Builder} builder
  * @param {string} entry
  * @param {string} dir
- * @param {import('./index.js').ServerlessConfig} config
- * @param {import('@sveltejs/kit').RouteDefinition[]} routes
+ * @param {import('.').ServerlessConfig} config
  */
-async function create_function_bundle(builder, entry, dir, config, routes) {
+async function create_function_bundle(builder, entry, dir, config) {
 	fs.rmSync(dir, { force: true, recursive: true });
 
 	let base = entry;
@@ -548,24 +605,6 @@ async function create_function_bundle(builder, entry, dir, config, routes) {
 			'\t'
 		)
 	);
-
-	const server_assets = builder.getServerAssets();
-	let routes_assets = new Set(server_assets.rootErrorPage);
-
-	for (const route of routes) {
-		const assets = server_assets.routes.get(route.id);
-		if (assets) {
-			routes_assets = new Set([...routes_assets, ...assets]);
-		}
-	}
-
-	if (server_assets.hooks) {
-		routes_assets = new Set([...routes_assets, ...server_assets.hooks]);
-	}
-
-	for (const asset of routes_assets) {
-		builder.copy(path.join(builder.getServerDirectory(), asset), path.join(dir, asset));
-	}
 
 	write(`${dir}/package.json`, JSON.stringify({ type: 'module' }));
 }