+++ packages/adapter-vercel/test/apps/split/src/hooks.server.js
@@ -1,6 +1,0 @@
-import text from '$lib/hooks.server.js.txt';
-
-export async function handle({ event, resolve }) {
-	event.setHeaders({ 'x-server-asset': text });
-	return resolve(event);
-}