+++ packages/adapter-vercel/test/apps/split/src/routes/load/+page.svelte
@@ -1,5 +1,0 @@
-<script>
-	import asset from '$lib/+page.svelte.txt';
-</script>
-
-<p>{asset}</p>