+++ packages/kit/test/apps/basics/test/cross-platform/client.test.js
@@ -240,11 +240,11 @@ test.describe('Navigation lifecycle functions', () => {
 
 	test('onNavigate calls callback', async ({ page, clicknav }) => {
 		await page.goto('/navigation-lifecycle/on-navigate/a');
-		expect(await page.textContent('h1')).toBe('undefined -> undefined (...)');
+		expect(await page.textContent('h1')).toBe('undefined -> undefined (...) false');
 
 		await clicknav('[href="/navigation-lifecycle/on-navigate/b"]');
 		expect(await page.textContent('h1')).toBe(
-			'/navigation-lifecycle/on-navigate/a -> /navigation-lifecycle/on-navigate/b (link)'
+			'/navigation-lifecycle/on-navigate/a -> /navigation-lifecycle/on-navigate/b (link) true'
 		);
 	});
 });
@@ -417,13 +417,29 @@ test.describe('Scrolling', () => {
 		await expect(page.locator('input')).toBeFocused();
 	});
 
-	test('scroll positions are recovered on reloading the page', async ({ page, app }) => {
+	test('scroll positions are recovered on reloading the page', async ({
+		page,
+		app,
+		browserName
+	}) => {
+		// No idea why the workaround below works only in dev mode
+		// A better solution would probably be to set fission.webContentIsolationStrategy: 1
+		// in the Firefox preferences but the Playwright API to do so is incomprehensible
+		if (!process.env.DEV && browserName === 'firefox') {
+			return;
+		}
+
 		await page.goto('/anchor');
 		await page.evaluate(() => window.scrollTo(0, 1000));
 		await app.goto('/anchor/anchor');
 		await page.evaluate(() => window.scrollTo(0, 1000));
 
 		await page.reload();
+		if (browserName === 'firefox') {
+			// Firefox with Playwright pushed new history entry history after reload
+			// See https://github.com/microsoft/playwright/issues/22640
+			await page.goBack();
+		}
 		expect(await page.evaluate(() => window.scrollY)).toBe(1000);
 
 		await page.goBack();